changing money to use bigdecimal

+++ src/org/capitalism/Product.java
@@ -9,7 +9,7 @@ public class Product implements IConsumable {
 	
 	public Product(Money money) {
 
-		value = money.getValue();
+		value = money.getValue().doubleValue();
 		priceMultiplier = new Random().nextDouble() + 1;
 		Money.MoneySupply.giveMoney(money);
 	}
